refactor(cli): status-command 포맷터를 status-formatters.js로 분리 (#50 step A)
status-command.js에서 format* 8개 함수를 cli/status-formatters.js로 이동.
status-command.js는 STATUS_COMMANDS / runStatusCommand / parseStatusOptions만
남긴다 (~45 LOC, 이전 226 LOC).

기존 import 경로 호환을 위해 status-command.js에서 전체 re-export.
테스트는 경로 변경 없이 그대로 통과.

doctor-command → doctor-helpers 패턴과 동일한 구조.

489 pass.

refs #50
