+++ packages/agent/src/cli/status-formatters.js
@@ -1,0 +1,176 @@
+/**
+ * status / usage 출력용 pure formatter 모음.
+ *
+ * 모든 함수는 string[] 반환, console.log 호출 없음.
+ * status-command.js의 runStatusCommand가 이 함수들의 결과를 출력한다.
+ */
+
+/**
+ * 전체 status 출력 라인 배열.
+ */
+export function formatStatusOutput(command, snapshot) {
+  const lines = [
+    `명령: ${command}`,
+    '로컬 에이전트 상태 요약',
+    '-----------------------',
+    `설정 파일: ${snapshot.configPath}`,
+    `Codex 사용: ${snapshot.providers.codex.enabled ? 'enabled' : 'disabled'}`,
+    `Claude 사용: ${snapshot.providers.claude.enabled ? 'enabled' : 'disabled'}`,
+    `서버 sync: ${snapshot.sync.enabled ? 'enabled' : 'disabled'}`,
+  ];
+  if (snapshot.accountFilter) {
+    lines.push(`계정 필터: ${snapshot.accountFilter}`);
+  }
+  lines.push(
+    '',
+    ...formatCodexSection(snapshot.codex),
+    '',
+    ...formatClaudeSection(snapshot.claude),
+  );
+  return lines;
+}
+
+/** Codex usage section. */
+export function formatCodexSection(codex) {
+  const lines = ['Codex usage', '-----------'];
+
+  if (!codex.enabled) {
+    lines.push('비활성화됨');
+    return lines;
+  }
+
+  lines.push(`인증 소스: ${codex.authSource ?? 'unknown'}`);
+  if (codex.authProfilesPath) {
+    lines.push(`Auth profiles 경로: ${codex.authProfilesPath}`);
+  }
+
+  if (codex.snapshots.length === 0) {
+    if (codex.filteredOut) {
+      lines.push(`계정 필터 "${codex.accountFilter}"에 해당하는 Codex 계정을 찾지 못했습니다.`);
+    } else {
+      lines.push('발견된 Codex OAuth 프로필이 없습니다.');
+    }
+    return lines;
+  }
+
+  for (const snapshot of codex.snapshots) {
+    const label = snapshot.account.email
+      ? `${snapshot.account.profileId} (${snapshot.account.email})`
+      : snapshot.account.profileId;
+    lines.push(`- ${label}`);
+    lines.push(
+      `  상태: ${
+        snapshot.status.ok
+          ? `OK (${snapshot.status.httpStatus})`
+          : `실패 (${snapshot.status.httpStatus ?? 'network/error'})`
+      }`,
+    );
+    lines.push(
+      `  source=${snapshot.source}, authType=${snapshot.authType}, confidence=${snapshot.confidence}`,
+    );
+    if (snapshot.account.plan) lines.push(`  플랜: ${snapshot.account.plan}`);
+    for (const window of snapshot.usageWindows) {
+      lines.push(`  ${window.kind}: ${formatWindow(window)}`);
+    }
+    if (snapshot.status.message) lines.push(`  에러: ${snapshot.status.message}`);
+  }
+  return lines;
+}
+
+/** Claude usage section. */
+export function formatClaudeSection(claude) {
+  const lines = ['Claude usage', '------------'];
+  lines.push(`인증 소스: ${claude.authSource}`);
+  lines.push(`credential 감지: ${claude.detected}`);
+  if (claude.selectedAccount && !claude.accountFilter) {
+    lines.push(`기본 계정: ${claude.selectedAccount.accountKey}`);
+  }
+
+  const usages = Array.isArray(claude.networkUsages)
+    ? claude.networkUsages
+    : claude.networkUsage
+      ? [{ accountKey: claude.selectedAccount?.accountKey ?? null, snapshot: claude.networkUsage }]
+      : [];
+  lines.push(
+    ...formatClaudeNetworkUsages(usages, {
+      filteredOut: claude.filteredOut,
+      accountFilter: claude.accountFilter,
+    }),
+  );
+  lines.push(...formatClaudeLocalUsage(claude.usage));
+  return lines;
+}
+
+/** Claude live network usage 블록(들). */
+export function formatClaudeNetworkUsages(usages, context = {}) {
+  const lines = ['', '[live] api.anthropic.com/api/oauth/usage'];
+  if (!usages || usages.length === 0) {
+    if (context.filteredOut) {
+      lines.push(`  계정 필터 "${context.accountFilter}"에 해당하는 Claude 계정을 찾지 못했습니다.`);
+    } else {
+      lines.push('  호출 안 함 (Claude 비활성 또는 토큰 없음)');
+    }
+    return lines;
+  }
+
+  for (const { accountKey, snapshot } of usages) {
+    if (usages.length > 1) lines.push(`  - 계정: ${accountKey ?? '(unknown)'}`);
+    lines.push(...formatClaudeNetworkUsageBody(snapshot, usages.length > 1));
+  }
+  return lines;
+}
+
+/** 단일 Claude network usage snapshot 출력. */
+export function formatClaudeNetworkUsageBody(networkUsage, indented = false) {
+  const prefix = indented ? '    ' : '  ';
+  const lines = [];
+  if (!networkUsage) {
+    lines.push(`${prefix}호출 안 함`);
+    return lines;
+  }
+
+  if (networkUsage.status?.ok) {
+    lines.push(`${prefix}상태: OK (${networkUsage.status.httpStatus})`);
+    if (networkUsage.usageWindows.length === 0) {
+      lines.push(`${prefix}usageWindows 없음 (응답에 기대한 필드가 없었음)`);
+    }
+    for (const window of networkUsage.usageWindows) {
+      lines.push(`${prefix}${window.kind}: ${formatWindow(window)}`);
+    }
+    return lines;
+  }
+
+  const http = networkUsage.status?.httpStatus ?? 'network/error';
+  const bucket = networkUsage.status?.bucket ?? 'unknown';
+  lines.push(`${prefix}상태: 실패 (${http}, bucket=${bucket})`);
+  if (networkUsage.status?.message) lines.push(`${prefix}메시지: ${networkUsage.status.message}`);
+  return lines;
+}
+
+/**
+ * @deprecated 신규 코드는 formatClaudeNetworkUsages(배열)를 사용.
+ */
+export function formatClaudeNetworkUsage(networkUsage) {
+  const header = ['', '[live] api.anthropic.com/api/oauth/usage'];
+  return [...header, ...formatClaudeNetworkUsageBody(networkUsage, false)];
+}
+
+/** Claude local stats-cache block. */
+export function formatClaudeLocalUsage(usage) {
+  const lines = ['', '[local] stats-cache.json'];
+  if (!usage || usage.source === 'not-found') {
+    lines.push('  데이터 없음 (stats-cache.json 미발견)');
+    return lines;
+  }
+  lines.push(`  총 세션 수: ${usage.totalSessions ?? '알 수 없음'}`);
+  lines.push(`  총 메시지 수: ${usage.totalMessages ?? '알 수 없음'}`);
+  lines.push(`  모델별 usage: ${usage.hasModelUsage ? '있음' : '없음'}`);
+  lines.push(`  일별 token 통계: ${usage.hasDailyModelTokens ? '있음' : '없음'}`);
+  return lines;
+}
+
+export function formatWindow(window) {
+  const reset = window.resetAt ? `reset_at=${window.resetAt}` : 'reset_at=unknown';
+  const used = window.usedPercent ?? 'unknown';
+  return `used_percent=${used}, ${reset}`;
+}
